// src/types/api.ts
export type SessionType = "reading" | "listening";
// ...
export interface NoteItem {
  noteId: string;
  sessionId: string;
  sessionType: SessionType;
  title: string;
  body: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface GetNotesResponse {
  items: NoteItem[];
}

export interface PostNoteRequest {
  sessionId: string;
  sessionType: SessionType;
  title: string;
  body: string;
  tags: string[];
}

export interface PostNoteResponse {
  item: NoteItem;
}

export interface PatchNoteRequest {
  title?: string;
  body?: string;
  tags?: string[];
}